ProbabilisticForecasting/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
arviz=0.19.0=pyhd8ed1ab_0
asttokens=2.0.5=pyhd3eb1b0_0
atk-1.0=2.38.0=h04ea711_2
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
bzip2=1.0.8=h4bc722e_7
c-ares=1.33.1=heb4867d_0
ca-certificates=2024.8.30=hbcca054_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.5.0=pyhd8ed1ab_0
cairo=1.18.0=hebfffa5_3
certifi=2024.8.30=py312h06a4308_0
cloudpickle=3.0.0=pyhd8ed1ab_0
comm=0.2.1=py312h06a4308_0
cons=0.4.6=pyhd8ed1ab_0
contourpy=1.3.0=py312h68727a3_1
cycler=0.12.1=pyhd8ed1ab_0
debugpy=1.6.7=py312h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
dm-tree=0.1.8=py312h83439f5_5
etuples=0.3.9=pyhd8ed1ab_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.6.3=h5888daf_0
filelock=3.16.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_2
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.54.0=py312h66e93f0_0
freetype=2.12.1=h267a509_2
fribidi=1.0.10=h36c2ea0_0
gdk-pixbuf=2.42.12=hb9ae30d_0
graphite2=1.3.13=h59595ed_1003
graphviz=12.0.0=hba01fac_0
gtk2=2.24.33=h6470451_5
gts=0.7.6=h977cf35_4
h5netcdf=1.3.0=pyhd8ed1ab_0
h5py=3.11.0=nompi_py312hb7ab980_102
harfbuzz=9.0.0=hda332d3_1
hdf5=1.14.3=nompi_hdf9ad27_105
icu=75.1=he02047a_0
ipykernel=6.28.0=py312h06a4308_0
ipython=8.27.0=py312h06a4308_0
ipywidgets=8.1.2=py312h06a4308_0
jedi=0.19.1=py312h06a4308_0
jupyter_client=8.6.0=py312h06a4308_0
jupyter_core=5.7.2=py312h06a4308_0
jupyterlab_widgets=3.0.10=py312h06a4308_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=py312h68727a3_0
krb5=1.21.3=h659f571_0
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.43=h712a8e2_0
lerc=4.0.0=h27087fc_0
libabseil=20240116.2=cxx17_he02047a_1
libaec=1.1.3=h59595ed_0
libblas=3.9.0=24_linux64_openblas
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=24_linux64_openblas
libcurl=8.10.1=hbbe4b11_0
libdeflate=1.21=h4bc722e_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=hd590300_2
libexpat=2.6.3=h5888daf_0
libffi=3.4.2=h7f98852_5
libgcc=14.1.0=h77fa898_1
libgcc-ng=14.1.0=h69a702a_1
libgd=2.3.3=hd3e95f3_10
libgfortran=14.1.0=h69a702a_1
libgfortran-ng=14.1.0=h69a702a_1
libgfortran5=14.1.0=hc5f4f2c_1
libglib=2.80.3=h315aac3_2
libgomp=14.1.0=h77fa898_1
libiconv=1.17=hd590300_2
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=24_linux64_openblas
libnghttp2=1.58.0=h47da74e_1
libnsl=2.0.1=hd590300_0
libopenblas=0.3.27=pthreads_hac2b453_1
libpng=1.6.44=hadc24fc_0
librsvg=2.58.4=hc0ffecb_0
libsodium=1.0.18=h7b6447c_0
libsqlite=3.46.1=hadc24fc_0
libssh2=1.11.0=h0841786_0
libstdcxx=14.1.0=hc0a3c3a_1
libstdcxx-ng=14.1.0=h4852527_1
libtiff=4.7.0=h6565414_0
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.16=hb9d3cd8_1
libxcrypt=4.4.36=hd590300_1
libxml2=2.12.7=he7c6b58_4
libzlib=1.3.1=h4ab18f5_1
logical-unification=0.4.6=pyhd8ed1ab_0
markdown-it-py=3.0.0=pyhd8ed1ab_0
matplotlib-base=3.9.2=py312hd3ec401_1
matplotlib-inline=0.1.6=py312h06a4308_0
mdurl=0.1.2=pyhd8ed1ab_0
minikanren=1.0.3=pyhd8ed1ab_0
multipledispatch=0.6.0=pyhd8ed1ab_1
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.5=he02047a_1
nest-asyncio=1.6.0=py312h06a4308_0
networkx=3.3=py312h06a4308_0
numpy=1.26.4=py312heda63a1_0
openjpeg=2.5.2=h488ebb8_0
openssl=3.3.2=hb9d3cd8_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.3=py312hf9745cd_1
pango=1.54.0=h4c5309f_1
parso=0.8.3=pyhd3eb1b0_0
pcre2=10.44=hba22ea6_2
pexpect=4.8.0=pyhd3eb1b0_3
pillow=10.4.0=py312h56024de_1
pip=24.2=pyh8b19718_1
pixman=0.43.2=h59595ed_0
platformdirs=3.10.0=py312h06a4308_0
prompt-toolkit=3.0.43=py312h06a4308_0
prompt_toolkit=3.0.43=hd3eb1b0_0
psutil=5.9.0=py312h5eee18b_0
pthread-stubs=0.4=hb9d3cd8_1002
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pygments=2.18.0=pyhd8ed1ab_0
pymc=5.16.2=hd8ed1ab_0
pymc-base=5.16.2=pyhd8ed1ab_0
pyparsing=3.1.4=pyhd8ed1ab_0
pytensor=2.23.0=py312h526ad5a_0
pytensor-base=2.25.4=py312h25a0e75_0
python=3.12.6=hc5c86c4_1_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-graphviz=0.20.3=pyh717bed2_0
python-tzdata=2024.2=pyhd8ed1ab_0
python_abi=3.12=5_cp312
pytz=2024.1=pyhd8ed1ab_0
pyzmq=25.1.2=py312h6a678d5_0
qhull=2020.2=h434a139_5
readline=8.2=h8228510_1
rich=13.8.1=pyhd8ed1ab_0
scipy=1.14.1=py312h7d485d2_0
setuptools=74.1.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
stack_data=0.2.0=pyhd3eb1b0_0
threadpoolctl=3.5.0=pyhc1e730c_0
tk=8.6.13=noxft_h4845f30_101
toolz=0.12.1=pyhd8ed1ab_0
tornado=6.4.1=py312h5eee18b_0
traitlets=5.14.3=py312h06a4308_0
typing-extensions=4.12.2=hd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h8827d51_1
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.44.0=pyhd8ed1ab_0
widgetsnbextension=4.0.10=py312h06a4308_0
xarray=2024.9.0=pyhd8ed1ab_0
xarray-einstats=0.8.0=pyhd8ed1ab_0
xorg-kbproto=1.0.7=hb9d3cd8_1003
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.9=hb711507_1
xorg-libxau=1.0.11=hb9d3cd8_1
xorg-libxdmcp=1.1.3=hb9d3cd8_1
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxrender=0.9.11=hd590300_0
xorg-renderproto=0.11.1=hb9d3cd8_1003
xorg-xextproto=7.3.0=hb9d3cd8_1004
xorg-xproto=7.0.31=hb9d3cd8_1008
xz=5.2.6=h166bdaf_0
zeromq=4.3.5=h6a678d5_0
zlib=1.3.1=h4ab18f5_1
zstd=1.5.6=ha6fb4c9_0
